debt: Moved action menu handling from FileExplorer to FileList
